Implement `Main()` to write a date.
`Main()` begins by registering components with the container. It
later calls `WriteDate()` to actually write the date. The method
`WriteDate()` creates a lifetime scope. Within that scope it asks the
container to resolve a concrete implementation of `IDateWriter`. The
container resolves that by instatiating a `TodayWriter`. To instantiate
this class requires an implementation of `IOutput`. The container
resolves that implementation to the concrete `ConsoleOutput` type.